删除敏感信息传递返回

+++ rainy_backend/src/main/java/com/rain/entity/common/CurrentUser.java
@@ -13,12 +13,12 @@ public class CurrentUser {
     //工厂模式，内部 实例化，无需外部实例化，设置为final
     private static  final ThreadLocal<Long> threadLocal=new ThreadLocal<>();
 
-    public static void setCurrentId(Long id){
-        threadLocal.set(id);
+    public static void setCurrentId(Integer id){
+        threadLocal.set(Long.valueOf(id));
     }
 
     //获取当前线程的ID
-    public static Long getCurrentId(){
-        return threadLocal.get();
+    public static Integer getCurrentId(){
+        return Math.toIntExact(threadLocal.get());
     }
 }
